BUGFIX !!!! Do you ever open up thousands of files in one directory?  It takes FOREVER.
I tested it by creating a directory with 10,000 files in it.  I thought maybe the data structure is slowing down as it fills up.  So I would expect it to start fast and then slowdown.
Nope.  I had it print the time every 100 files.  It took about 2 seconds for every 100 files.  Which is 200 (!!!) seconds for 10000 files!  Same time spent for every 100 files.

Next I went through the code in the 10,000 item loop.  There was this call:

long time = fs->Timestamp();

It fetched the timestamp for every file.  I left that in the code all this time for no reason.  I never used the timestamp!  This looked super-suspicious.  So I deleted that one line.

Result -- it read in 10,000 files in under 2 seconds!  A 100X performance improvement!

getting the timestamp was taking 99+% of the time